Fix warning about .CRT-section present. Might help with #70.
Not all of these changes might be necessary. I am no expert when it comes to c++ initialization rules, but all static members and global variables have been eliminated.

+++ libblankenhain/src/BhMath.cpp
@@ -81,7 +81,7 @@ namespace BhMath {
 	// for convenience, we implement pow as |x|^y
 	float pow(float x, float y)
 	{
-		x = abs(x) + 1e-8;
+		x = abs(x) + static_cast<float>(1e-8);
 		float result = 0.f;
 		__asm
 		{
